+++ packages/version/index.js
@@ -1,5 +1,5 @@
 module.exports = {
-  "decorator": "3.20.0",
-  "core": "3.20.0",
-  "timestamp": 1738332637497
+  "decorator": "3.20.3",
+  "core": "3.20.3",
+  "timestamp": 1740920210221
 };