+++ scripts/rollback/rollback-3.20.3.sh
@@ -1,0 +1,132 @@
+#!/bin/bash
+# timestamp: 1740920033709
+
+npm dist-tag add @midwayjs/fc-starter@3.20.0 latest
+npm dist-tag add @midwayjs/async-hooks-context-manager@3.20.0 latest
+npm dist-tag add @midwayjs/axios@3.20.2 latest
+npm dist-tag add @midwayjs/bootstrap@3.20.0 latest
+npm dist-tag add @midwayjs/bull@3.20.0 latest
+npm dist-tag add @midwayjs/bull-board@3.20.2 latest
+npm dist-tag add @midwayjs/bullmq@3.20.2 latest
+npm dist-tag add @midwayjs/busboy@3.20.2 latest
+npm dist-tag add @midwayjs/cache-manager@3.20.0 latest
+npm dist-tag add @midwayjs/captcha@3.20.2 latest
+npm dist-tag add @midwayjs/casbin@3.20.2 latest
+npm dist-tag add @midwayjs/casbin-redis-adapter@3.20.2 latest
+npm dist-tag add @midwayjs/casbin-typeorm-adapter@3.20.2 latest
+npm dist-tag add @midwayjs/code-dye@3.20.2 latest
+npm dist-tag add @midwayjs/consul@3.20.2 latest
+npm dist-tag add @midwayjs/core@3.20.0 latest
+npm dist-tag add @midwayjs/cos@3.20.0 latest
+npm dist-tag add @midwayjs/cron@3.20.0 latest
+npm dist-tag add @midwayjs/cross-domain@3.20.2 latest
+npm dist-tag add @midwayjs/decorator@3.20.0 latest
+npm dist-tag add @midwayjs/etcd@3.20.0 latest
+npm dist-tag add @midwayjs/express-session@3.20.0 latest
+npm dist-tag add @midwayjs/faas@3.20.0 latest
+npm dist-tag add @midwayjs/grpc@3.20.2 latest
+npm dist-tag add @midwayjs/http-proxy@3.20.2 latest
+npm dist-tag add @midwayjs/i18n@3.20.2 latest
+npm dist-tag add @midwayjs/info@3.20.2 latest
+npm dist-tag add @midwayjs/jwt@3.20.2 latest
+npm dist-tag add @midwayjs/kafka@3.20.2 latest
+npm dist-tag add @midwayjs/leoric@3.20.2 latest
+npm dist-tag add @midwayjs/mikro@3.20.2 latest
+npm dist-tag add @midwayjs/mock@3.20.0 latest
+npm dist-tag add @midwayjs/mongoose@3.20.1 latest
+npm dist-tag add @midwayjs/mqtt@3.20.2 latest
+npm dist-tag add @midwayjs/oss@3.20.0 latest
+npm dist-tag add @midwayjs/otel@3.20.2 latest
+npm dist-tag add @midwayjs/passport@3.20.2 latest
+npm dist-tag add @midwayjs/process-agent@3.20.2 latest
+npm dist-tag add @midwayjs/prometheus@3.20.2 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@3.20.2 latest
+npm dist-tag add @midwayjs/rabbitmq@3.20.2 latest
+npm dist-tag add @midwayjs/redis@3.20.0 latest
+npm dist-tag add @midwayjs/security@3.20.2 latest
+npm dist-tag add @midwayjs/sequelize@3.20.2 latest
+npm dist-tag add @midwayjs/session@3.20.0 latest
+npm dist-tag add @midwayjs/socketio@3.20.2 latest
+npm dist-tag add @midwayjs/static-file@3.20.2 latest
+npm dist-tag add @midwayjs/swagger@3.20.2 latest
+npm dist-tag add @midwayjs/tablestore@3.20.0 latest
+npm dist-tag add @midwayjs/tags@3.20.2 latest
+npm dist-tag add @midwayjs/tenant@3.20.0 latest
+npm dist-tag add @midwayjs/typegoose@3.20.1 latest
+npm dist-tag add @midwayjs/typeorm@3.20.0 latest
+npm dist-tag add @midwayjs/upload@3.20.2 latest
+npm dist-tag add @midwayjs/validate@3.20.2 latest
+npm dist-tag add @midwayjs/version@3.20.2 latest
+npm dist-tag add @midwayjs/view@3.20.2 latest
+npm dist-tag add @midwayjs/view-ejs@3.20.2 latest
+npm dist-tag add @midwayjs/view-nunjucks@3.20.2 latest
+npm dist-tag add @midwayjs/web@3.20.2 latest
+npm dist-tag add @midwayjs/express@3.20.2 latest
+npm dist-tag add @midwayjs/koa@3.20.2 latest
+npm dist-tag add @midwayjs/ws@3.20.2 latest
+
+# Changes:
+
+#  - @midwayjs/fc-starter: 3.20.0 => 3.20.3
+#  - @midwayjs/async-hooks-context-manager: 3.20.0 => 3.20.3
+#  - @midwayjs/axios: 3.20.2 => 3.20.3
+#  - @midwayjs/bootstrap: 3.20.0 => 3.20.3
+#  - @midwayjs/bull: 3.20.0 => 3.20.3
+#  - @midwayjs/bull-board: 3.20.2 => 3.20.3
+#  - @midwayjs/bullmq: 3.20.2 => 3.20.3
+#  - @midwayjs/busboy: 3.20.2 => 3.20.3
+#  - @midwayjs/cache-manager: 3.20.0 => 3.20.3
+#  - @midwayjs/captcha: 3.20.2 => 3.20.3
+#  - @midwayjs/casbin: 3.20.2 => 3.20.3
+#  - @midwayjs/casbin-redis-adapter: 3.20.2 => 3.20.3
+#  - @midwayjs/casbin-typeorm-adapter: 3.20.2 => 3.20.3
+#  - @midwayjs/code-dye: 3.20.2 => 3.20.3
+#  - @midwayjs/consul: 3.20.2 => 3.20.3
+#  - @midwayjs/core: 3.20.0 => 3.20.3
+#  - @midwayjs/cos: 3.20.0 => 3.20.3
+#  - @midwayjs/cron: 3.20.0 => 3.20.3
+#  - @midwayjs/cross-domain: 3.20.2 => 3.20.3
+#  - @midwayjs/decorator: 3.20.0 => 3.20.3
+#  - @midwayjs/etcd: 3.20.0 => 3.20.3
+#  - @midwayjs/express-session: 3.20.0 => 3.20.3
+#  - @midwayjs/faas: 3.20.0 => 3.20.3
+#  - @midwayjs/grpc: 3.20.2 => 3.20.3
+#  - @midwayjs/http-proxy: 3.20.2 => 3.20.3
+#  - @midwayjs/i18n: 3.20.2 => 3.20.3
+#  - @midwayjs/info: 3.20.2 => 3.20.3
+#  - @midwayjs/jwt: 3.20.2 => 3.20.3
+#  - @midwayjs/kafka: 3.20.2 => 3.20.3
+#  - @midwayjs/leoric: 3.20.2 => 3.20.3
+#  - @midwayjs/mikro: 3.20.2 => 3.20.3
+#  - @midwayjs/mock: 3.20.0 => 3.20.3
+#  - @midwayjs/mongoose: 3.20.1 => 3.20.3
+#  - @midwayjs/mqtt: 3.20.2 => 3.20.3
+#  - @midwayjs/oss: 3.20.0 => 3.20.3
+#  - @midwayjs/otel: 3.20.2 => 3.20.3
+#  - @midwayjs/passport: 3.20.2 => 3.20.3
+#  - @midwayjs/process-agent: 3.20.2 => 3.20.3
+#  - @midwayjs/prometheus: 3.20.2 => 3.20.3
+#  - @midwayjs/prometheus-socket-io: 3.20.2 => 3.20.3
+#  - @midwayjs/rabbitmq: 3.20.2 => 3.20.3
+#  - @midwayjs/redis: 3.20.0 => 3.20.3
+#  - @midwayjs/security: 3.20.2 => 3.20.3
+#  - @midwayjs/sequelize: 3.20.2 => 3.20.3
+#  - @midwayjs/session: 3.20.0 => 3.20.3
+#  - @midwayjs/socketio: 3.20.2 => 3.20.3
+#  - @midwayjs/static-file: 3.20.2 => 3.20.3
+#  - @midwayjs/swagger: 3.20.2 => 3.20.3
+#  - @midwayjs/tablestore: 3.20.0 => 3.20.3
+#  - @midwayjs/tags: 3.20.2 => 3.20.3
+#  - @midwayjs/tenant: 3.20.0 => 3.20.3
+#  - @midwayjs/typegoose: 3.20.1 => 3.20.3
+#  - @midwayjs/typeorm: 3.20.0 => 3.20.3
+#  - @midwayjs/upload: 3.20.2 => 3.20.3
+#  - @midwayjs/validate: 3.20.2 => 3.20.3
+#  - @midwayjs/version: 3.20.2 => 3.20.3
+#  - @midwayjs/view: 3.20.2 => 3.20.3
+#  - @midwayjs/view-ejs: 3.20.2 => 3.20.3
+#  - @midwayjs/view-nunjucks: 3.20.2 => 3.20.3
+#  - @midwayjs/web: 3.20.2 => 3.20.3
+#  - @midwayjs/express: 3.20.2 => 3.20.3
+#  - @midwayjs/koa: 3.20.2 => 3.20.3
+#  - @midwayjs/ws: 3.20.2 => 3.20.3